requirements.txt
Flask==3.1.2
Flask_SocketIO==5.5.1
insightface==0.7.3
numpy==2.3.2
opencv_python==4.11.0.86
Pillow==11.3.0
psycopg2==2.9.10
python-dotenv==1.1.1
Requests==2.32.5
torch==2.8.0.dev20250517+cu128
ultralytics==8.3.152
